Use composer to arrange code and the plugin itself for wp-env
