--- app/autoload.php
@@ -1,17 +1,0 @@
-<?php
-    require_once("models/table_entry.php");
-    require_once("models/sams_table.php");
-    require_once("presenters/table_entry_presenter.php");
-    require_once("presenters/table_presenter.php");
-    require_once("remote/uris/table_uri.php");
-    require_once("remote/xml_fetcher.php");
-    require_once("remote/file_cacher.php");
-    require_once("sams_table_renderer.php");
-
-    require_once("models/fixtures_entry.php");
-    require_once("models/sams_fixtures.php");
-    require_once("presenters/fixtures_entry_presenter.php");
-    require_once("presenters/fixtures_presenter.php");
-    require_once("remote/uris/fixtures_uri.php");
-    require_once("sams_fixtures_renderer.php");
-?>
