+++ includes/SAMSFixturesRenderer.php
