+++ includes/templates/__fixtures_entry.html
