+++ includes/templates/__table_entry.html
